move mongo-code to new module
